Added new 'ManagedFeedImage' and 'ManagedCache' to separate files for organization, and added the 'managedImages', 'localFeedImages' and 'localFeed' function helpers to make the code on production more readable
