+++ FeedStoreChallenge/CoreDataFeedStore.swift
@@ -34,7 +34,7 @@ public final class CoreDataFeedStore: FeedStore {
 				let request = NSFetchRequest<ManagedCache>(entityName: ManagedCache.entity().name!)
 				request.returnsObjectsAsFaults = false
 				if let cache = try context.fetch(request).first {
-					completion(.found(feed: cache.feed.compactMap { ($0 as? ManagedFeedImage) }.map { LocalFeedImage(id: $0.id, description: $0.imageDescription, location: $0.location, url: $0.url) }, timestamp: cache.timestamp))
+					completion(.found(feed: cache.localFeed, timestamp: cache.timestamp))
 				} else {
 					completion(.empty)
 				}
@@ -48,15 +48,8 @@ public final class CoreDataFeedStore: FeedStore {
 		perform { context in
 			do {
 				let managedCache = ManagedCache(context: context)
+				managedCache.feed = ManagedFeedImage.managedImages(from: feed, in: context)
 				managedCache.timestamp = timestamp
-				managedCache.feed = NSOrderedSet(array: feed.map { local in
-					let managed = ManagedFeedImage(context: context)
-					managed.id = local.id
-					managed.imageDescription = local.description
-					managed.location = local.location
-					managed.url = local.url
-					return managed
-				})
 
 				try context.save()
 				completion(nil)
@@ -74,19 +67,4 @@ public final class CoreDataFeedStore: FeedStore {
 		let context = self.context
 		context.perform { action(context) }
 	}
-}
-
-@objc(ManagedCache)
-private class ManagedCache: NSManagedObject {
-	@NSManaged var timestamp: Date
-	@NSManaged var feed: NSOrderedSet
-}
-
-@objc(ManagedFeedImage)
-private class ManagedFeedImage: NSManagedObject {
-	@NSManaged var id: UUID
-	@NSManaged var imageDescription: String?
-	@NSManaged var location: String?
-	@NSManaged var url: URL
-	@NSManaged var cache: ManagedCache
 }
