+++ FeedStoreChallenge/ManagedCache.swift
@@ -1,0 +1,19 @@
+//
+//  ManagedCache.swift
+//  FeedStoreChallenge
+//
+//  Created by Nicolás Cadena on 17/04/21.
+//  Copyright © 2021 Essential Developer. All rights reserved.
+//
+
+import CoreData
+
+@objc(ManagedCache)
+internal class ManagedCache: NSManagedObject {
+	@NSManaged var timestamp: Date
+	@NSManaged var feed: NSOrderedSet
+
+	var localFeed: [LocalFeedImage] {
+		return feed.compactMap { ($0 as? ManagedFeedImage)?.localFeedImage }
+	}
+}
